+++ sickbeard/providers/generic.py
@@ -56,7 +56,7 @@ class GenericProvider:
 
         self.search_mode = None
         self.search_fallback = False
-        self.enable_daily = False
+        self.enable_recentsearch = False
         self.enable_backlog = False
 
         self.cache = tvcache.TVCache(self)